pyproject.toml
[tool.poetry]
name = "sql-filler"
version = "0.1.0"
description = ""
authors = ["sanikk <sanikk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
invoke = "^2.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
